requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
pydantic
python-jose[cryptography]
httpx
aiofiles
